refactor identifier to account and added account APIs
