+++ src/main/java/org/mifos/connector/gsma/HealthCheck.java
@@ -3,22 +3,12 @@ package org.mifos.connector.gsma;
 import org.apache.camel.Exchange;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.jackson.JacksonDataFormat;
-import org.apache.camel.model.dataformat.JsonLibrary;
 import org.mifos.connector.gsma.auth.dto.AccessTokenStore;
-import org.mifos.connector.gsma.identifier.dto.AccountStatusRequest;
-import org.mifos.connector.gsma.transfer.dto.GSMATransaction;
 import org.mifos.connector.gsma.zeebe.ZeebeProcessStarter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
-
-import static org.mifos.connector.gsma.camel.config.CamelProperties.*;
 
 @Component
 public class HealthCheck extends RouteBuilder {
