+++ src/main/java/org/mifos/connector/gsma/account/AccountWorkers.java
@@ -1,10 +1,13 @@
-package org.mifos.connector.gsma.identifier;
+package org.mifos.connector.gsma.account;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.zeebe.client.ZeebeClient;
 import org.apache.camel.CamelContext;
 import org.apache.camel.Exchange;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.support.DefaultExchange;
+import org.mifos.connector.common.channel.dto.TransactionChannelRequestDTO;
+import org.mifos.connector.common.mojaloop.dto.PartyIdInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +21,7 @@ import static org.mifos.connector.gsma.camel.config.CamelProperties.*;
 import static org.mifos.connector.gsma.zeebe.ZeebeExpressionVariables.PAYEE_LOOKUP_RETRY_COUNT;
 
 @Component
-public class IdentifierLookupWorkers {
+public class AccountWorkers {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
@@ -34,6 +37,8 @@ public class IdentifierLookupWorkers {
     @Value("${zeebe.client.evenly-allocated-max-jobs}")
     private int workerMaxJobs;
 
+    @Autowired
+    private ObjectMapper objectMapper;
 
     @PostConstruct
     public void setupWorkers() {
@@ -50,6 +55,14 @@ public class IdentifierLookupWorkers {
                     exchange.setProperty(CHANNEL_REQUEST, variables.get("channelRequest"));
                     exchange.setProperty(ACCOUNT_ACTION, "status");
                     exchange.setProperty(IS_RTP_REQUEST, variables.get(IS_RTP_REQUEST));
+
+                    TransactionChannelRequestDTO channelRequest = objectMapper.readValue(exchange.getProperty(CHANNEL_REQUEST, String.class), TransactionChannelRequestDTO.class);
+                    PartyIdInfo requestedParty = exchange.getProperty(IS_RTP_REQUEST, Boolean.class) ? channelRequest.getPayer().getPartyIdInfo() : channelRequest.getPayee().getPartyIdInfo();
+
+                    exchange.setProperty(IDENTIFIER_TYPE, requestedParty.getPartyIdType().toString().toLowerCase());
+                    exchange.setProperty(IDENTIFIER, requestedParty.getPartyIdentifier());
+
+                    exchange.setProperty(IS_API_CALL, "false");
 
                     producerTemplate.send("direct:account-route", exchange);
